+++ src/index.js
@@ -1,39 +1,14 @@
 /* eslint-disable */
 import '@fortawesome/fontawesome-free/js/all.js';
 import './style.css';
+import storageAvailable from '../module/storageAvailable';
 
-// Check for availaible storage
-const storageAvailable = (type) => {
-    let storage;
-    try {
-      storage = window[type];
-      const x = '__storage_test__';
-      storage.setItem(x, x);
-      storage.removeItem(x);
-      return true;
-    } catch (e) {
-      return (
-        e instanceof DOMException
-        // everything except Firefox
-        && (e.code === 22
-          // Firefox
-          || e.code === 1014
-          // test name field too, because code might not be present
-          // everything except Firefox
-          || e.name === 'QuotaExceededError'
-          // Firefox
-          || e.name === 'NS_ERROR_DOM_QUOTA_REACHED')
-        // acknowledge QuotaExceededError only if there's something already stored
-        && storage
-        && storage.length !== 0
-      );
-    }
-  };
+
+
 
 const form = document.querySelector("form");
 const cont = document.querySelector(".small-container");
 // const inputbtn =document.querySelector('.listItems');
-const inputfield = document.querySelector(".class-input");
 
 
 class ListToDo{
@@ -61,15 +36,15 @@ function addListTask(description){
     console.log(description)
     return newTask;
 }
-// function removeStorage(){
-//     getListTask();
-//     const listId = +e.target.id.slice(5);
-//     const filteredList = taskList.filter((task)=>task.index !==listId);
-//     for (let i = 0; i < filteredList.length; i +=1){
-//         filteredList[i].index = i +1
-//     }
-//     saveTask(filteredList)
-// }
+function removeStorage(){
+    getListTask();
+    const listId = +e.target.id.slice(5);
+    const filteredList = listarr.filter((task)=>task.index !==listId);
+    for (let i = 0; i < filteredList.length; i +=1){
+        filteredList[i].index = i +1
+    }
+    saveTask(filteredList)
+}
 // ----------------------end storage-----------------------
 
 
@@ -81,7 +56,7 @@ div.innerHTML = `
 <input type="text" class="class-input" value="${chores.description}" readonly>
 </div>
 <div>
-<button class="edit">edit</button><button class="del-btn"> del</button> <button class="fa-solid fa-ellipsis-vertical">del</button> </div>
+<button class="edit">edit</button><button class="del-btn"> del</button> <button id="btn-id-${chores.index} class="fa-solid fa-ellipsis-vertical">del</button> </div>
 
 
 `;
@@ -128,10 +103,14 @@ if(storageAvailable('localStorage')){
     })
     form.addEventListener("submit", function(e){
         e.preventDefault();
-        const input = document.getElementById("todo").value;
+
+       
+            const input = document.getElementById("todo").value;
         const anotherTask = addListTask(input);
         displayitems(anotherTask)
         saveStorage();
+        
+        
     })
     
 }
